cli: address review comments and better align with conventions per revciew
