+++ controller/pkg/cli/root.go
@@ -9,7 +9,7 @@ import (
 	controllercmd "github.com/agentgateway/agentgateway/controller/pkg/cli/controller"
 	"github.com/agentgateway/agentgateway/controller/pkg/cli/costs"
 	"github.com/agentgateway/agentgateway/controller/pkg/cli/flag"
-	migrationcmd "github.com/agentgateway/agentgateway/controller/pkg/cli/migration"
+	"github.com/agentgateway/agentgateway/controller/pkg/cli/migrate"
 	proxycmd "github.com/agentgateway/agentgateway/controller/pkg/cli/proxy"
 	"github.com/agentgateway/agentgateway/controller/pkg/cli/trace"
 	cliversion "github.com/agentgateway/agentgateway/controller/pkg/cli/version"
@@ -26,7 +26,7 @@ func NewRootCmd() *cobra.Command {
 	rootCmd.AddCommand(proxycmd.Command())
 	rootCmd.AddCommand(controllercmd.Command())
 	rootCmd.AddCommand(costs.Command())
-	rootCmd.AddCommand(migrationcmd.Command())
+	rootCmd.AddCommand(migrate.Command())
 
 	rootCmd.AddCommand(flag.BuildCobra(config.Command))
 	rootCmd.AddCommand(flag.BuildCobra(trace.Command))
